refactor: renamed deploy to release
